added cart items

+++ app/datatables/user_datatable.rb
@@ -25,12 +25,14 @@ class UserDatatable < AjaxDatatablesRails::ActiveRecord
         admin: record.admin ? 'Yes' : 'No',
         edit: link_to(
           heroicon('pencil', variant: :outline,
-                             options: { class: 'hover:text-indigo-700 hover:cursor-pointer h-4' }), edit_user_path(record.user_id)
+                             options: { class: 'hover:text-indigo-700 hover:cursor-pointer h-4' }),
+          edit_user_path(record.user_id)
         ),
         destroy: link_to(
           heroicon('trash', variant: :outline,
-                            options: { class: 'hover:text-indigo-700 hover:cursor-pointer h-4' }), user_path(record.user_id),
-          method: :delete
+                            options: { class: 'hover:text-indigo-700 hover:cursor-pointer h-4' }),
+          user_path(record.user_id),
+          method: :delete, data: { confirm: 'Are you sure you want to proceed?' }
         )
       }
     end
